Renamed sorting algo container
