+++ apps/desktop/applications/AlgorithmVisualizer/Algorithms/BubbleSort.tsx
@@ -1,7 +1,7 @@
 
 import { SubViewParams } from "../AlgorithmVisualizerView";
 import { SortView } from "./SortingView";
-import { SortingAlgorithmContainer } from "./AlgorithmContainer";
+import { SortingAlgorithmContainer } from "./SortingAlgorithmContainer";
 
 async function bubbleSort(view: SortView, abortSignal: AbortSignal) {
   for (let i = 0; i < view.size() - 1; i++) {
