+++ apps/desktop/applications/AlgorithmVisualizer/Algorithms/SortingAlgorithmContainer.module.css
